code refactor - options files is moved to common folder inside services
