+++ src/client/services/common/optionsData.js
@@ -1,4 +1,8 @@
 export const identity = [
+    {
+		value: "",
+		gender: "Please Select",
+	},
 	{
 		value: "M",
 		gender: "Male",
@@ -46,6 +50,7 @@ export const results = [
 		result: " "
 	}
 ];
+
 export const resultsSearch = [
 	{
 		code: "all",
@@ -85,6 +90,220 @@ export const faxTypes =[
 	{
 		value: "ManualFax",
 		faxType: "ManualFax",
+	},
+];
+
+export const ethnicity = [
+	{
+		value: "",
+		desc: "Please Select",
+	},
+	{
+		value: "Hispanic or Latino",
+		desc: "Hispanic or Latino",
+	},
+	{
+		value: "Not Hispanic or Latino",
+		desc: "Not Hispanic or Latino",
+	},
+];
+
+export const race = [
+	{
+		id: 1,
+		value: "White",
+	},
+	{
+		id: 2,
+		value: "Black or African American",
+	},
+	{
+		id: 3,
+		value: "American Indian or Alsaka Native",
+	},
+	{
+		id: 4,
+		value: "Asian",
+	},
+	{
+		id: 5,
+		value: "Native Hawaiian or Other Pacific Islander",
+	},
+	{
+		id: 6,
+		value: "Other",
+	},
+];
+
+export const symptoms = [
+	{
+		id: 1,
+		value: "None",
+	},
+	{
+		id: 2,
+		value: "Contact with and (suspected) exposure",
+	},
+	{
+		id: 3,
+		value: "Fever or chills",
+	},
+	{
+		id: 4,
+		value: "Cough",
+	},
+	{
+		id: 5,
+		value: "Shortness of breath / difficulty breathing",
+	},
+	{
+		id: 6,
+		value: "Fatigue",
+	},
+	{
+		id: 7,
+		value: "Muscle / body aches",
+	},
+	{
+		id: 8,
+		value: "Loss of taste or smell",
+	},
+	{
+		id: 9,
+		value: "Sore throat",
+	},
+	{
+		id: 10,
+		value: "Congestion or runny nose",
+	},
+	{
+		id: 11,
+		value: "Nausea or vomiting",
+	},
+];
+
+export const insuranceProvider = [
+	{
+		id:1,
+		value: "",
+		desc:"Please Select"
+	},
+	{
+		id:2,
+		value: "Self/Pay",
+		desc: "Self/Pay",
+	},
+	{
+		id: 3,
+		value: "Aetna",
+		desc: "Aetna",
+	},
+	{
+		id: 4,
+		value: "Anthem Blue Cross Blue Shield",
+		desc: "Anthem Blue Cross Blue Shield",
+	},
+	{
+		id: 5,
+		value: "Blue Shield Of California",
+		desc: "Blue Shield Of California",
+	},
+	{
+		id: 6,
+		value: "CareMore",
+		desc: "CareMore",
+	},
+	{
+		id: 7,
+		value: "HealthNet",
+		desc: "HealthNet",
+	},
+	{
+		id: 8,
+		value: "Humana",
+		desc: "Humana",
+	},
+	{
+		id: 9,
+		value: "Inland Empire Health Plan",
+		desc: "Inland Empire Health Plan",
+	},
+	{
+		id: 10,
+		value: "Kaiser Permanente",
+		desc: "Kaiser Permanente",
+	},
+	{
+		id: 11,
+		value: "Medicare",
+		desc: "Medicare",
+	},
+	{
+		id: 12,
+		value: "Molina Healthcare",
+		desc: "Molina Healthcare",
+	},
+	{
+		id: 13,
+		value: "Oxford (UnitedHealthCare)",
+		desc: "Oxford (UnitedHealthCare)",
+	},
+	{
+		id: 14,
+		value: "SAG-AFTRA Health Plan",
+		desc: "SAG-AFTRA Health Plan",
+	},
+
+	{
+		id: 15,
+		value: "SCAN Health Plan",
+		desc: "SCAN Health Plan",
+	},
+	{
+		id: 16,
+		value: "State of California (Medi-Cal)",
+		desc: "State of California (Medi-Cal)",
+	},
+	{
+		id: 17,
+		value: "United Healthcare",
+		desc: "United Healthcare",
+	},
+	{
+		id: 18,
+		value: "Other",
+		desc: "Other",
+	},
+];
+
+export const relation = [
+	{
+		value: "",
+		desc: "Please Select",
+	},
+	{
+		value: "Self",
+		desc: "Self",
+	},
+	{
+		value: "Spouse",
+		desc: "Spouse",
+	},
+	{
+		value: "Parent",
+		desc: "Parent",
+	},
+	{
+		value: "Child",
+		desc: "Child",
+	},
+	{
+		value: "Other Adult",
+		desc: "Other Adult",
+	},
+	{
+		value: "Adult",
+		desc: "Adult",
 	},
 ];
